.pipeline/projects/fiverr_job_automation_tool/workspace/requirements.txt
# Fiverr Job Automation Tool Dependencies

# HTTP client for API requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Testing framework
pytest>=7.4.0
pytest-mock>=3.12.0

# Type hints support
typing-extensions>=4.8.0
